Split drawing code from model-examining code
Remove all examination of models from the ComportexViz client code and
move it into a server namespace that can run equally well in the
browser or on a physical server. I'll change some of these from cljs
to cljc in the near future.

Abandon the proxy approach. Rather than sending a minimal amount of
model data to the client, just have the server tell the client what to
draw.

I pulled each existing draw function apart into a data function and a
draw function. So this might prove to be an "initial refactor". Now
that they're split apart, there might be cleaner ways of organizing
the pieces. I apologize for the `cell-segments-data` function.

Concepts introduced:
- The "simulation" and the "journal". ComportexViz always connects to
   a journal (a recorded set of steps) and usually connects to a
   simulation (the ability to generate more steps)
- The "step-template". Get everybody out of the "wait for the next
  model step to see the latest specs" business. This finally
  eliminates the :on-model-changed message and simplifies the params
  tab, which now simply edits the step-template.
- The "selection" now contains a "model-id". Now when the selected
  model changes to due an incoming step, the "selection" changes,
  making it a good candidate for add-watch (reagent and regular)