Visualize Comportex while it runs on the JVM
Design decision: Use Jetty's websockets. Don't hide them under
core.async, just use them directly. Don't use a library that attempts
to solve websockets for Clojure and ClojureScript, just use the raw
API in the browser. I think there'd be lots of wrestling with
libraries, e.g. in configuring Transit write-handlers and
read-handlers, and it just seems best to work closer to the core APIs
until we find a good reason not to.

Created a way of sending channels over the network via "channel
proxies". Note that this is not the same as treating the websocket
itself as a channel.

Got rid of "sim-options". It was too hard to synchronize an atom
between two machines without causing ping-ponging. Do everything via
into-sim and subscribing to play/pause status.