state transition diagram.
A first attempt. Matches active (learnable) cells to previously active sets of cells, i.e. SDRs. Shows predicted sequence transitions between these SDRs derived from connected distal synapses. This display is useful even just for distinguishing states in terms of cells, rather than the columns shown in the main display.

Currently the display shows distinct states in an order of first appearance. That is nice in that it is easy to interpet, but will not scale up well, and does not show similarity between SDRs. A force-directed graph layout might work as an option. The order-of-appearance one will need a scroll ability too.

The transitions-plot component has to keep track of the active cells each time step to match/update the SDRs it stores. That should maybe be moved to the server side to avoid a lot of network traffic.

This plot component -- like the time series plots -- needs to update its state each time step. But we should not have to actually draw it if the tab is not visible. So need to separate these concerns.