Stop rendering background tabs.
This is a significant [low-hanging fruit] perf win for rendering.

Gather and aggregate the time series plot data independently of the
plots.

Now that the cell excitation plot doesn't render in the background,
there's no need to hide it behind a checkbox.

Use a resizing-canvas for the ts-freqs-plot. Tweak the plot-size to
use the whole canvas.

Fix a bug in canvas and resizing-canvas. They don't call the
most-recently-passed-in draw function, which is a problem when the
function refers to non-atom values via closures. Reagent makes it
really hard to render a canvas correctly.

I stared at the aggregation code for a while and came up with an
approach that uses only immutable values. Also, support non-keyword
keys in aggregate-by fn, though I didn't end up taking advantage of
this.