Comportex Notebook, minus blogging
I ported Comportex Notebook to this new "talk-to-a-server" model and
brought it into ComportexViz.

I created a "hello.clj" worksheet. It introduces the Comportex API,
creating timelines in the notebook, creating graphs, and starting a
runner.

The visualizations will only work when connected to the server that
holds the underlying models. So you generally should clear all output
before saving a worksheet.