Specify the viz-canvas size externally
Take the resizing-canvas apart into two pieces: the part that adapts
to the element's size and the part that notifies of size changes.
Create a window-resize-listener empty component which takes advantage
of React's lifecycle methods for listening and unlistening to events.

Adopt a philosophy on how a component should avoid screwing with
parent-owned channels. On unmount, a component must immediately stop
taking values off of the channel. The only ways to do this are (a)
make the parent pass in a mult, in which case the child doesn't hurt
anything by swiping values, or (b) make the child keep a teardown
channel / mult, and guard all takes from parent-owned channels using
alt! or alts!! with priority true, prioritizing the teardown channel.
I switched from (a) to (b) because I like parents more than children.