Clean up the connection persistence
Rather than a client-side :connection-admin target, have a server-side
:connection-persistor. Use a subscribe approach for the reconnect
blob.

Rather than making the client keep a list of remote targets that
receive the :client-reconnect message, just send it to the
:connection-persistor and require the server to take it from there. On
the server side, route this via a connection-changes channel, tapping
the relevant channels into it.

Instate journal-ids and simulation-ids in the client-info so that one
connection can support multiple journals, etc.