Make inputs clickable, show their synapses
Useful for analyzing which columns are receiving input from a given
bit. Especially if you show inactive feed-forward synapses.

Because the selection is no longer only region -> layer -> column and
is now often input -> inbit, I changed the selection to use a "path"
and a "bit". So the word "bit" refers to either the column or the
input bit. This "path" corresponds to the layout path.

Now that the selection holds a layout path, lots of code is
simplified. See "scroll!", "add-facets!", etc.

Currently the inputs are sortable, scrollable, and facet-able. This
might be silly. Or it might be desirable. I haven't decided. Easy to
change, just pass in layer-only paths.