Timeline selection improvements
Include the timeline in notebook viz.

Make command-clicking the timeline actually useful. It duplicates the
most recent selection, just on a new timestep, so you can quickly
compare, e.g., the set of connected synapses on multiple timesteps.

Grab the actual timesteps from each model. In the notebook it's
convenient to use take-nth to sample a series of steps. So the
timeline can no longer assume consecutive timesteps.