Make the viz-canvas grow to fit its contents
Rather than sizing the viz-canvas via CSS, let it size itself, and
impose limits via a max-height and max-width in the viz-options.

This is important for the notebook, where whitespace is gross.

This means we now resize this canvas via script, not CSS. With CSS it
would be difficult for the component to know whether it should
maintain its current size. Now there's a single pair of authoritative
numbers.

This required calculating the cells-segments layout before render,
similar to the other layouts. So the cells-segments layout now lives
inside viz-layouts.

I removed awareness of the source layout, bit, and dt from the
cells-segments layout. The method that uses those now takes them as
params. This matches other nearby usage, e.g. drawing distal synapses
from segments to other regions, whereas it's awkward to grab this
information at my new layout calculation point.

Also:

- Fix a bug I recently put in the viz-timeline. Now it listens to
  window resize.
- Change some occurrences of "cell-segs" to "cells-segs" in hopes of
  preventing future pain.
- Don't draw breaks in 1D.