Allow partial rows in coerced-to-2D visualizations
Changed the Grid2dLayout to make it aware that sometimes
non-rectangles are coerced into rectangles. Changed all layouts to
expose an "ids-count", and changed the viz-canvas to always use this
value rather than using the topology's size.

Fixed off-by-one errors in "id-onscreen?". I tried to make the code
more obviously right by no longer subtracting one, and instead doing
an "and" of "<=" and "<".