Remove a landmine. main/comportexviz-app is now allowed to rerender.
Fixes #9

Adopt a pattern: never pass a raw component function into another
component. Put it in a vector in the caller, not the callee. This
gives the caller the ability to pass in arguments. Previously I hacked
in this ability by having the caller put it in a vector and then wrap
that in an anonymous function. The gaping flaw in this approach was
that a new function was created on each parent render, so React had no
way of knowing whether the child component was of the same type, since
the function object identifies the type of component, and the only way
for two functions to be equivalent is if they're the same object. So
React would unmount the old component and mount a new one.