Move some text out of the canvas and into the DOM
Via React, it'll be easy to change the viz-canvas into a mashup of
canvas, divs, SVG, etc. I moved the top labels out of the canvas and
into the DOM. The benefit of this is that browsers are way better at
rendering text in the DOM than in canvases. It's especially noticeable
on Retina screens or when using pinch-to-zoom.

I think it'd be good to move all text labels out of the canvas, and
also write the viz-timeline in SVG (which has the same benefits, but
for shapes as well).

This change just handles the top labels, though, and updates the
notebook to keep it compatible.