Draw boxes around message parameters: Channels
This change rethinks channel proxies.

The essential parts of the channel proxy are:

- During decoding, information is added to the channel. Namely, a function for
  sending messages across the network.
- It's a channel. Code that expects a channel will just work, whether the
  channel is typical or if it points to another machine.

This change takes the first idea further. Now we also add information to the
channel *during serialization*. Rather than registering a ChannelProxy for each
message, just draw a box around the channel (a 'ChannelMarshal'). Then when it's
serialized, register it. This has the huge benefit of removing 'local-targets'
from being plumbed everywhere. This registry still exists, but it only needs to
sit near the networking code.

The change fits this into a new vocabulary: marshalling. I think this idea of
drawing boxes around message parameters can go further. Another box ('marshal')
I want to create is a 'ShouldCache' type. You just wrap it around a value, and
the networking code does magic to only send it across the network as necessary.
In fact, it was imagining the 'ShouldCache' that made me realize this channel
change was possible. Anyway, maybe soon. I think it's a useful concept.

I had to do some reconnect-blob rethinking. Hence the 'marshal/channel-weak'.
I think the reconnection logic is still pretty tedious.

I thought about just adding Transit write handlers for raw channels, but boxing
them gives other opportunities like putting a `release!` method on the box, and
it just feels cleaner to express intent.

On that note, this change also fixes a big leak problem that I forgot I'd
created. The local-targets and their corresponding channels are now released
when we're done with them.

In the journal and simulation, stop doing so much inside of go-blocks. I was
running into some nasty Clojure (or cljs or core.async?) bugs with destructuring
inside of go-blocks. StackOverflowErrors with no explanation. It was
bad. Instead of rage-quitting, I decided to stop stress-testing go blocks. I
moved the command handling out into a function. This is better anyway, you can
set browser breakpoints on the code now that it's not being mangled by macros.

Bump to latest versions of clj, cljs, transit, core.async. I did this to try
solving the go-block issue. It didn't work, but it didn't hurt anything.