Draw boxes around message parameters: Big Values
Make the messaging/network layer more magical.

I'm happy so much of this logic fits in marshalling.cljc.

I quickly changed the viewport to use this. Now it just sends the value without
having to register it first. No more tokens. But this is really just a
placeholder. The idea of the 'viewport' code can now be rethought, now that it's
much more lightweight to send different large values to the server.

It's unclear how often we'll need to mark a value as large, but one inherently
large value is 'ids-onscreen', which has to be large because of sorting.