Unbundle :get-cells-segments into smaller commands
Use three commands. One to get cells. One for segments. And one for synapses.
And have separate distal and apical commands, so it's five total.

This removes knowledge of selected cells or selected segments from the server.
Now, upon receiving the segments data, the client chooses which synapses to
request synapses for.

I'm happy to finally kill off the 'cells-segments-data' function on the server,
and to stop storing a cells-segs-response on the client. There's still room for
more improvement.

I like this direction. The server no longer knows that cells, segments, and
synapses are all displayed together. Good for flexibility, and good for simpler
servers.