Send strings, not keywords, across the network
The main goal for this change is to make it where servers aren't inconvenienced
by the fact that Sanity is written in ClojureScript.

The overall approach: Clojure endpoints (e.g. the Sanity client and the
Comportex server) will use keywords internally. They use stringify-keys and
keywordize-keys, so all map keys just work. Some manual switching between
keywords and strings does happen because there are values that appear as both
keys and message params (namely, region IDs and layer IDs).

This change required me to form a point of view on keywords. I'll blog about it
here. :)

Clojure untangles a "keyword" type from the "string" type. The initial reason
for having a dedicated "keyword" is that it can do magic to support super-fast
equality checks. With keywords, a central instance table is used, so that two
equal-valued keywords are guaranteed to use the same instance.

In Clojure, equality checks are super fast if the references are equal (=>
values equal). They're pretty fast when the references are different and the
cached hashes are different (=> values not equal), just requiring a one-time
hash check per instance. Equality checks are much slower when the values are the
same but the references are non-equal. Then equality checking requires comparing
the entire value every time.

For keywords, you only need the reference check.

During map lookup, the key's hash is used to reach a leaf node which contains a
list of keys with colliding hashes, and then the key is compared to each of
these keys to check for a match. So, fast equality checking is a good
thing. This kind of thing is possible because keywords are untangled from
strings, so this central instance table overhead doesn't apply to strings.

So the above explains why keywords are a good idea in general: for perf. But why
should we use them on Sanity endpoints built in Clojure? Will it actually be
good for performance, since it involves translating every sent and received
message (with 'stringify-keys' and 'keywordize-keys')?

Two answers:

- Our reason for keywordizing the strings isn't for perf. It's because the
  Clojure language itself embraces keywords. Clojure favors keywords over
  strings as map keys for perf, and the language then adds additional reasons to
  use keywords (e.g. map destructuring). We want keywords for those reasons.
- We are free to fix this perf issue. We could use custom Transit handlers for
  Maps which translate the keys between keywords and strings. I didn't do this,
  but we have this in our back pocket.

Other notes:

- Similar to my other client-server changes, the in-browser Comportex is largely
  unaffected. It doesn't do the stringify + keywordize.
- Similar to my other client-server changes, this is the initial big
  change. With this new line drawn, over time we might improve some of this
  keyword/string back and forth, e.g. with region IDs
- I banged my head against the wall repeatedly while chasing down an issue that
  ended up being a Transit bug.