Add a browser REPL connector in the UI
I've only tested it with {:optimizations :none}. And the experience isn't
anything special. And I didn't bother with autoreloads or anything like that,
because I'd have to figure out how to make the autoreloads include all the
source locations (including demos).

I needed this REPL to reduce the Transit bug:
https://github.com/cognitect/transit-cljs/issues/22