Replace "get-inbits-cols" with "get-sense-bits" and "get-layer-bits"
Now we query for the bits one layer / sense at a time. This makes the
message-passing simpler. The server doesn't have to communicate which layer /
sense this data is for. So values like :regions or :layer-3 don't have to get
sent, so we don't have to solve whether they should be strings or keywords.