Fetch "active-columns" via messages, not from the step
Solve the draw-to :all ff-synapse problem differently. Instead, just re-fetch
the proximal segments when we're in this scenario.

Putting this on the step is messy. It caused issues when I wanted to keywordize
/ stringify on a message-by-message basis, since part of the step should be
keywordized while others shouldn't be (layer IDs, region IDs).

So this goes back to drawing steps as they come in, often before their bits are
available.