Only keywordize when it makes sense
This was my initial approach on the keyword->string change, but I ran into
issues the first time, so I switched to the universal keywordize /
stringify. But that's a bad approach because of the way it mangles keys that are
themselves data, e.g. IDs.

My recent bits / steps changes addressed the issues I ran into.

In general, we keywordize when storing the value for later use. In general, the
server just uses strings. And we never mangle region IDs, sense IDs, or layer
IDs, so they're always the expected type. Some servers will use keyword IDs,
others will use strings, and the client will take whatever it's given.