Perf improvements
Faster "which steps are new" and "which steps are old" inference. Use an
into-viz message to communicate dropped steps. Assume new steps are at the
beginning of the list. So in both cases, no more scanning all the steps.

Also speed up a keywordize and stop triggering layout rebuilds on behalf of
distal/apical segments when a bit isn't selected.