Split "subscribe" into 3 messages. Fix and improve model resetting.
Also rename the "step template" to the "network shape". The previous name
reflected my old design where each step got merged with its fetched data.

We now keywordize parts of the network shape when it arrives. Add this treatment
to the demos, unbreaking resets.

Also make the viz-canvas stop using the up-to-date network shape when fetching
bits. Instead, assoc the network shape onto the step so that we don't do
anything goofy after adding more regions and then selecting an old
step. Regardless of how important resetting is, this is elegant. Note that we
still do goofy things with segments if you select an old bit in a newly created
region.