The "seg-selector": making synapse / segment messages more powerful
Sometimes we want to query for synapses on multiple columns, and it's slow to do
it one column at a time. The API ought to allow for batch queries.

Instead of patching batch queries on top, e.g. by making messages take in a list
of columns, this change solves the general problem. Fundamentally, we already
were serving segments / synapses in batches, just in a hardcoded way -- by
column, or by segment. This change gets rid of the hardcode. The sender can now
create advanced column / cell / segment selectors to specify what it wants.

I like that the API no longer feels like it was designed around Sanity's UI.

In cases where we query lots of columns / segments, using these batch queries is
a nice perf win, bringing us roughly back up to where we were when the server
did all the fetch logic.