New message: "get-layer-stats". Replaces "get-column-state-freqs".
Now we fetch for one layer at a time.

This new message handler doesn't call core/column-state-freqs. It seemed easier
to fetch data from the layer directly rather than use a layer-fn. We could
always create a 'core/layer-column-state-freqs' function, but I don't know if
Comportex really needs to be in the business of providing a suite of helpers.

I didn't put this inside "get-layer-bits" as another "fetch" because it would be
ambiguous whether the onscreen-bits should affect the results.