Fetch "cells-per-column" via one message
Change the selection so that it remembers the step, not just the
snapshot-id. This gives access to the selected step's network shape, which gives
access to the number of cells per columns.

Also add a fetches parameter to "get-column-cells".

Also fix a bug where I forgot to fetch the capture options in main.cljs.